requirements.txt
aiohttp==3.9.5
discord.py==2.2.2
imgkit==1.2.3
Jinja2==3.1.5
Pillow==11.1.0
PyMySQL==1.0.2
python-dotenv==1.0.1
Requests==2.32.3
pytest==8.3.5
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-asyncio==0.23.8
black==25.1.0
vulture==2.14
scikit-learn==1.5.0
numpy==1.26.3
docker
